Secure Cloud Storage App
A cloud service that allows users to securely upload text file and
images. All uploaded files are encrypted before moved in storage, and
the encryption key is not saved. When a user uploads a file, they must
type a Hint and Pass-phrase specific for that uploaded file. If their
responses pass input validation, the file is encrypted using a hashed
(MD5) randomly generated number as the key (utilized 256-bit
encryption). This key is then encrypted using the users inputted
Pass-phrase hashed (again, MD5). Both the hash version, and plaintext
version of the pass-phrase are not saved anywhere. Instead, when the
user wishes to download or delete an uploaded file, they must type the
pass-phrase for that file (the hint inputted when they uploaded the file
will be displayed ), which will be hashed and used to decrypt the file
key. If the entered pass-phrase is incorrect, the key to the file, and
hence, the file, will not decrypted correctly and won't be deleted or
downloaded.

More a in-depth description of the app such as directory security, as
well as the secuirty threats I activiley protected against can be read
in my Thesis, linked below.
http://www.mediafire.com/view/de56rxu3ktr38ry/Quizi_Secure_Cloud_Computing_Thesis.docx